add ability for client-side authority on !collect results
If LocationAutoCollect is true, the AP server will be the source of truth for whether a location has been collected.
If LocationAutoCollect is false, the save file will be the source of truth for whether a location has been collected.
LocationAutoCollect defaults true and the player has not yet been given a way to change that setting.